Rename record to timeline and remove old timeline
