+++ Cargo.toml
@@ -13,7 +13,6 @@ exclude = ["/.travis.yml"]
 edition = "2021"
 
 [dependencies]
-arrayvec = { version = "0.7", default-features = false }
 chrono = { version = "0.4", optional = true }
 colored = { version = "2", optional = true }
 serde = { version = "1", optional = true, default-features = false, features = ["derive"] }
@@ -21,7 +20,7 @@ serde = { version = "1", optional = true, default-features = false, features = [
 [features]
 default = ["alloc", "colored"]
 alloc = ["serde?/alloc"]
-serde = ["dep:serde", "chrono?/serde", "arrayvec/serde"]
+serde = ["dep:serde", "chrono?/serde"]
 
 [badges]
 maintenance = { status = "actively-developed" }
