+++ src/timeline.rs
@@ -1,28 +1,34 @@
 //! A timeline of actions.
 
-use crate::slot::{NoOp, Signal, Slot, SW};
-use crate::{Action, Entry, Merged, Result};
-use arrayvec::ArrayVec;
-use core::fmt;
+use crate::entry::{Entries, Stack};
+use crate::slot::{NoOp, Slot, SW};
+use crate::{Action, At, Entry, Format, History, Result};
+use alloc::{
+    collections::VecDeque,
+    string::{String, ToString},
+    vec::Vec,
+};
+use core::ops::{Index, IndexMut};
+use core::{
+    fmt::{self, Write},
+    num::NonZeroUsize,
+};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
-#[cfg(feature = "alloc")]
-use {
-    crate::{At, Format},
-    alloc::string::{String, ToString},
-    core::fmt::Write,
-};
 #[cfg(feature = "chrono")]
 use {
     chrono::{DateTime, Utc},
+    core::cmp::Ordering,
     core::convert::identity,
 };
 
-/// A timeline of actions.
+/// A linear timeline of actions.
 ///
-/// A timeline works mostly like a record but stores a fixed number of actions on the stack.
-///
-/// Can be used without the `alloc` feature.
+/// The timeline can roll the targets state backwards and forwards by using
+/// the undo and redo methods. In addition, the timeline can notify the user
+/// about changes to the stack or the target through [signal](enum.Signal.html).
+/// The user can give the timeline a function that is called each time the state
+/// changes by using the [`builder`](struct.TimelineBuilder.html).
 ///
 /// # Examples
 /// ```
@@ -30,7 +36,7 @@ use {
 /// # include!("../add.rs");
 /// # fn main() {
 /// let mut target = String::new();
-/// let mut timeline = Timeline::<_, _, 32>::new();
+/// let mut timeline = Timeline::new();
 /// timeline.apply(&mut target, Add('a')).unwrap();
 /// timeline.apply(&mut target, Add('b')).unwrap();
 /// timeline.apply(&mut target, Add('c')).unwrap();
@@ -47,124 +53,113 @@ use {
 /// ```
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Clone, Debug)]
-pub struct Timeline<A, F, const LIMIT: usize> {
-    entries: ArrayVec<Entry<A>, LIMIT>,
-    current: usize,
-    saved: Option<usize>,
-    slot: SW<F>,
+pub struct Timeline<A, F = NoOp> {
+    pub(crate) stack: Stack<LimitDeque<A>, F>,
 }
 
-impl<A, const LIMIT: usize> Timeline<A, NoOp, LIMIT> {
+impl<A> Timeline<A> {
     /// Returns a new timeline.
-    pub fn new() -> Timeline<A, NoOp, LIMIT> {
-        Timeline {
-            entries: ArrayVec::new(),
-            current: 0,
-            saved: Some(0),
-            slot: SW::default(),
-        }
+    pub fn new() -> Timeline<A> {
+        TimelineBuilder::new().build()
     }
 }
 
-impl<A, F, const LIMIT: usize> Timeline<A, F, LIMIT> {
+impl<A, F> Timeline<A, F> {
+    /// Returns a new builder for a timeline.
+    pub fn builder() -> TimelineBuilder<F> {
+        TimelineBuilder::new()
+    }
+
+    /// Reserves capacity for at least `additional` more actions.
+    ///
+    /// # Panics
+    /// Panics if the new capacity overflows usize.
+    pub fn reserve(&mut self, additional: usize) {
+        self.stack.entries.deque.reserve(additional);
+    }
+
+    /// Returns the capacity of the timeline.
+    pub fn capacity(&self) -> usize {
+        self.stack.entries.deque.capacity()
+    }
+
+    /// Shrinks the capacity of the timeline as much as possible.
+    pub fn shrink_to_fit(&mut self) {
+        self.stack.entries.deque.shrink_to_fit();
+    }
+
     /// Returns the number of actions in the timeline.
     pub fn len(&self) -> usize {
-        self.entries.len()
+        self.stack.entries.deque.len()
     }
 
     /// Returns `true` if the timeline is empty.
     pub fn is_empty(&self) -> bool {
-        self.entries.is_empty()
+        self.stack.entries.deque.is_empty()
     }
 
-    /// Returns the limit of the record.
-    pub const fn limit(&self) -> usize {
-        LIMIT
+    /// Returns the limit of the timeline.
+    pub fn limit(&self) -> usize {
+        self.stack.entries.limit.get()
     }
 
     /// Sets how the signal should be handled when the state changes.
-    ///
-    /// The previous slot is returned if it exists.
     pub fn connect(&mut self, slot: F) -> Option<F> {
-        self.slot.connect(Some(slot))
+        self.stack.slot.connect(Some(slot))
     }
 
     /// Removes and returns the slot if it exists.
     pub fn disconnect(&mut self) -> Option<F> {
-        self.slot.disconnect()
+        self.stack.slot.disconnect()
     }
 
     /// Returns `true` if the timeline can undo.
     pub fn can_undo(&self) -> bool {
-        self.current() > 0
+        self.stack.can_undo()
     }
 
     /// Returns `true` if the timeline can redo.
     pub fn can_redo(&self) -> bool {
-        self.current() < self.len()
+        self.stack.can_redo()
     }
 
     /// Returns `true` if the target is in a saved state, `false` otherwise.
     pub fn is_saved(&self) -> bool {
-        self.saved.map_or(false, |saved| saved == self.current())
+        self.stack.is_saved()
     }
 
     /// Returns the position of the current action.
     pub fn current(&self) -> usize {
-        self.current
+        self.stack.current
     }
 
-    /// Returns a structure for configurable formatting of the record.
-    #[cfg(feature = "alloc")]
-    pub fn display(&self) -> Display<A, F, LIMIT> {
+    /// Returns a queue.
+    pub fn queue(&mut self) -> Queue<A, F> {
+        Queue::from(self)
+    }
+
+    /// Returns a checkpoint.
+    pub fn checkpoint(&mut self) -> Checkpoint<A, F> {
+        Checkpoint::from(self)
+    }
+
+    /// Returns a structure for configurable formatting of the timeline.
+    pub fn display(&self) -> Display<A, F> {
         Display::from(self)
     }
 }
 
-impl<A: Action, F: Slot, const LIMIT: usize> Timeline<A, F, LIMIT> {
+impl<A: Action, F: Slot> Timeline<A, F> {
     /// Pushes the action on top of the timeline and executes its [`apply`] method.
     ///
     /// # Errors
     /// If an error occur when executing [`apply`] the error is returned.
     ///
     /// [`apply`]: trait.Action.html#tymethod.apply
-    pub fn apply(&mut self, target: &mut A::Target, mut action: A) -> Result<A> {
-        let output = action.apply(target)?;
-        let current = self.current();
-        let could_undo = self.can_undo();
-        let could_redo = self.can_redo();
-        let was_saved = self.is_saved();
-        // Pop off all elements after len from record.
-        self.entries.truncate(current);
-        // Check if the saved state was popped off.
-        self.saved = self.saved.filter(|&saved| saved <= current);
-        // Try to merge actions unless the target is in a saved state.
-        let merged = match self.entries.last_mut() {
-            Some(last) if !was_saved => last.action.merge(action),
-            _ => Merged::No(action),
-        };
-        match merged {
-            Merged::Yes => (),
-            Merged::Annul => {
-                self.entries.pop();
-                self.current -= 1;
-            }
-            // If actions are not merged or annulled push it onto the record.
-            Merged::No(action) => {
-                // If limit is reached, pop off the first action.
-                if LIMIT == self.current() {
-                    self.entries.pop_at(0);
-                    self.saved = self.saved.and_then(|saved| saved.checked_sub(1));
-                } else {
-                    self.current += 1;
-                }
-                self.entries.push(Entry::from(action));
-            }
-        };
-        self.slot.emit_if(could_redo, Signal::Redo(false));
-        self.slot.emit_if(!could_undo, Signal::Undo(true));
-        self.slot.emit_if(was_saved, Signal::Saved(false));
-        Ok(output)
+    pub fn apply(&mut self, target: &mut A::Target, action: A) -> Result<A> {
+        self.stack
+            .apply(target, action)
+            .map(|(output, _, _)| output)
     }
 
     /// Calls the [`undo`] method for the active action and sets
@@ -175,18 +170,7 @@ impl<A: Action, F: Slot, const LIMIT: usize> Timeline<A, F, LIMIT> {
     ///
     /// [`undo`]: ../trait.Action.html#tymethod.undo
     pub fn undo(&mut self, target: &mut A::Target) -> Option<Result<A>> {
-        self.can_undo().then(|| {
-            let was_saved = self.is_saved();
-            let old = self.current();
-            let output = self.entries[self.current - 1].action.undo(target)?;
-            self.current -= 1;
-            let is_saved = self.is_saved();
-            self.slot.emit_if(old == self.len(), Signal::Redo(true));
-            self.slot.emit_if(old == 1, Signal::Undo(false));
-            self.slot
-                .emit_if(was_saved != is_saved, Signal::Saved(is_saved));
-            Ok(output)
-        })
+        self.stack.undo(target)
     }
 
     /// Calls the [`redo`] method for the active action and sets
@@ -197,49 +181,24 @@ impl<A: Action, F: Slot, const LIMIT: usize> Timeline<A, F, LIMIT> {
     ///
     /// [`redo`]: trait.Action.html#method.redo
     pub fn redo(&mut self, target: &mut A::Target) -> Option<Result<A>> {
-        self.can_redo().then(|| {
-            let was_saved = self.is_saved();
-            let old = self.current();
-            let output = self.entries[self.current].action.redo(target)?;
-            self.current += 1;
-            let is_saved = self.is_saved();
-            self.slot
-                .emit_if(old == self.len() - 1, Signal::Redo(false));
-            self.slot.emit_if(old == 0, Signal::Undo(true));
-            self.slot
-                .emit_if(was_saved != is_saved, Signal::Saved(is_saved));
-            Ok(output)
-        })
+        self.stack.redo(target)
     }
 
     /// Marks the target as currently being in a saved or unsaved state.
     pub fn set_saved(&mut self, saved: bool) {
-        let was_saved = self.is_saved();
-        if saved {
-            self.saved = Some(self.current());
-            self.slot.emit_if(!was_saved, Signal::Saved(true));
-        } else {
-            self.saved = None;
-            self.slot.emit_if(was_saved, Signal::Saved(false));
-        }
+        self.stack.set_saved(saved)
     }
 
     /// Removes all actions from the timeline without undoing them.
     pub fn clear(&mut self) {
-        let could_undo = self.can_undo();
-        let could_redo = self.can_redo();
-        self.entries.clear();
-        self.saved = self.is_saved().then_some(0);
-        self.current = 0;
-        self.slot.emit_if(could_undo, Signal::Undo(false));
-        self.slot.emit_if(could_redo, Signal::Redo(false));
+        self.stack.clear()
     }
 }
 
-impl<A: Action<Output = ()>, F: Slot, const LIMIT: usize> Timeline<A, F, LIMIT> {
+impl<A: Action<Output = ()>, F: Slot> Timeline<A, F> {
     /// Revert the changes done to the target since the saved state.
     pub fn revert(&mut self, target: &mut A::Target) -> Option<Result<A>> {
-        self.saved.and_then(|saved| self.go_to(target, saved))
+        self.stack.revert(target)
     }
 
     /// Repeatedly calls [`undo`] or [`redo`] until the action at `current` is reached.
@@ -250,133 +209,351 @@ impl<A: Action<Output = ()>, F: Slot, const LIMIT: usize> Timeline<A, F, LIMIT> 
     /// [`undo`]: trait.Action.html#tymethod.undo
     /// [`redo`]: trait.Action.html#method.redo
     pub fn go_to(&mut self, target: &mut A::Target, current: usize) -> Option<Result<A>> {
-        if current > self.len() {
-            return None;
-        }
-        let could_undo = self.can_undo();
-        let could_redo = self.can_redo();
-        let was_saved = self.is_saved();
-        // Temporarily remove slot so they are not called each iteration.
-        let slot = self.disconnect();
-        // Decide if we need to undo or redo to reach current.
-        let f = if current > self.current() {
-            Timeline::redo
-        } else {
-            Timeline::undo
-        };
-        while self.current() != current {
-            if let Err(err) = f(self, target).unwrap() {
-                self.slot.connect(slot);
-                return Some(Err(err));
-            }
-        }
-        // Add slot back.
-        self.slot.connect(slot);
-        let can_undo = self.can_undo();
-        let can_redo = self.can_redo();
-        let is_saved = self.is_saved();
-        self.slot
-            .emit_if(could_undo != can_undo, Signal::Undo(can_undo));
-        self.slot
-            .emit_if(could_redo != can_redo, Signal::Redo(can_redo));
-        self.slot
-            .emit_if(was_saved != is_saved, Signal::Saved(is_saved));
-        Some(Ok(()))
+        self.stack.go_to(target, current)
     }
 
-    /// Go back or forward in the record to the action that was made closest to the datetime provided.
+    /// Go back or forward in the timeline to the action that was made closest to the datetime provided.
     #[cfg(feature = "chrono")]
     pub fn time_travel(&mut self, target: &mut A::Target, to: &DateTime<Utc>) -> Option<Result<A>> {
-        let current = self
-            .entries
-            .binary_search_by(|e| e.timestamp.cmp(to))
-            .unwrap_or_else(identity);
+        let current = match self.stack.entries.deque.as_slices() {
+            ([], []) => return None,
+            (head, []) => head
+                .binary_search_by(|e| e.timestamp.cmp(to))
+                .unwrap_or_else(identity),
+            ([], tail) => tail
+                .binary_search_by(|e| e.timestamp.cmp(to))
+                .unwrap_or_else(identity),
+            (head, tail) => match head.last().unwrap().timestamp.cmp(to) {
+                Ordering::Less => head
+                    .binary_search_by(|e| e.timestamp.cmp(to))
+                    .unwrap_or_else(identity),
+                Ordering::Equal => head.len(),
+                Ordering::Greater => {
+                    head.len()
+                        + tail
+                            .binary_search_by(|e| e.timestamp.cmp(to))
+                            .unwrap_or_else(identity)
+                }
+            },
+        };
         self.go_to(target, current)
     }
 }
 
-#[cfg(feature = "alloc")]
-impl<A: ToString, F, const LIMIT: usize> Timeline<A, F, LIMIT> {
+impl<A: ToString, F> Timeline<A, F> {
     /// Returns the string of the action which will be undone
     /// in the next call to [`undo`](struct.Timeline.html#method.undo).
     pub fn undo_text(&self) -> Option<String> {
-        self.current.checked_sub(1).and_then(|i| self.text(i))
+        self.stack.current.checked_sub(1).and_then(|i| self.text(i))
     }
 
     /// Returns the string of the action which will be redone
     /// in the next call to [`redo`](struct.Timeline.html#method.redo).
     pub fn redo_text(&self) -> Option<String> {
-        self.text(self.current)
+        self.text(self.stack.current)
     }
 
     fn text(&self, i: usize) -> Option<String> {
-        self.entries.get(i).map(|e| e.action.to_string())
+        self.stack
+            .entries
+            .deque
+            .get(i)
+            .map(|e| e.action.to_string())
     }
 }
 
-impl<A, const LIMIT: usize> Default for Timeline<A, NoOp, LIMIT> {
-    fn default() -> Timeline<A, NoOp, LIMIT> {
+impl<A> Default for Timeline<A> {
+    fn default() -> Timeline<A> {
         Timeline::new()
     }
 }
 
-/// Builder for a Timeline.
+impl<A, F> From<History<A, F>> for Timeline<A, F> {
+    fn from(history: History<A, F>) -> Timeline<A, F> {
+        history.timeline
+    }
+}
+
+/// Builder for a timeline.
+///
+/// # Examples
+/// ```
+/// # include!("../add.rs");
+/// # fn main() {
+/// # use undo::{timeline::TimelineBuilder, Timeline};
+///
+/// let _ = TimelineBuilder::new()
+///     .limit(100)
+///     .capacity(100)
+///     .connect(|s| { dbg!(s); })
+///     .build::<Add>();
+/// # }
+/// ```
 #[derive(Debug)]
-pub struct Builder<F = NoOp> {
+pub struct TimelineBuilder<F = NoOp> {
+    capacity: usize,
+    limit: NonZeroUsize,
     saved: bool,
     slot: SW<F>,
 }
 
-impl<F> Builder<F> {
-    /// Returns a builder for a record.
-    pub fn new() -> Builder<F> {
-        Builder {
+impl<F> TimelineBuilder<F> {
+    /// Returns a builder for a timeline.
+    pub fn new() -> TimelineBuilder<F> {
+        TimelineBuilder {
+            capacity: 0,
+            limit: NonZeroUsize::new(usize::MAX).unwrap(),
             saved: true,
             slot: SW::default(),
         }
     }
 
+    /// Sets the capacity for the timeline.
+    pub fn capacity(mut self, capacity: usize) -> TimelineBuilder<F> {
+        self.capacity = capacity;
+        self
+    }
+
+    /// Sets the `limit` of the timeline.
+    ///
+    /// # Panics
+    /// Panics if `limit` is `0`.
+    pub fn limit(mut self, limit: usize) -> TimelineBuilder<F> {
+        self.limit = NonZeroUsize::new(limit).expect("limit can not be `0`");
+        self
+    }
+
     /// Sets if the target is initially in a saved state.
     /// By default the target is in a saved state.
-    pub fn saved(mut self, saved: bool) -> Builder<F> {
+    pub fn saved(mut self, saved: bool) -> TimelineBuilder<F> {
         self.saved = saved;
         self
     }
 
-    /// Builds the record.
-    pub fn build<A, const LIMIT: usize>(self) -> Timeline<A, F, LIMIT> {
+    /// Builds the timeline.
+    pub fn build<A>(self) -> Timeline<A, F> {
         Timeline {
-            entries: ArrayVec::new(),
-            current: 0,
-            saved: self.saved.then_some(0),
-            slot: self.slot,
+            stack: Stack {
+                entries: LimitDeque {
+                    deque: VecDeque::with_capacity(self.capacity),
+                    limit: self.limit,
+                },
+                current: 0,
+                saved: self.saved.then_some(0),
+                slot: self.slot,
+            },
         }
     }
 }
 
-impl<F: Slot> Builder<F> {
+impl<F: Slot> TimelineBuilder<F> {
     /// Connects the slot.
-    pub fn connect(mut self, f: F) -> Builder<F> {
+    pub fn connect(mut self, f: F) -> TimelineBuilder<F> {
         self.slot = SW::new(f);
         self
     }
 }
 
-impl Default for Builder {
+impl Default for TimelineBuilder {
     fn default() -> Self {
-        Builder::new()
+        TimelineBuilder::new()
+    }
+}
+
+#[derive(Debug)]
+enum QueueAction<A> {
+    Apply(A),
+    Undo,
+    Redo,
+}
+
+/// Wraps a timeline and gives it batch queue functionality.
+///
+/// # Examples
+/// ```
+/// # use undo::{Timeline};
+/// # include!("../add.rs");
+/// # fn main() {
+/// let mut string = String::new();
+/// let mut timeline = Timeline::new();
+/// let mut queue = timeline.queue();
+/// queue.apply(Add('a'));
+/// queue.apply(Add('b'));
+/// queue.apply(Add('c'));
+/// assert_eq!(string, "");
+/// queue.commit(&mut string).unwrap().unwrap();
+/// assert_eq!(string, "abc");
+/// # }
+/// ```
+#[derive(Debug)]
+pub struct Queue<'a, A, F> {
+    timeline: &'a mut Timeline<A, F>,
+    actions: Vec<QueueAction<A>>,
+}
+
+impl<A: Action<Output = ()>, F: Slot> Queue<'_, A, F> {
+    /// Queues an `apply` action.
+    pub fn apply(&mut self, action: A) {
+        self.actions.push(QueueAction::Apply(action));
+    }
+
+    /// Queues an `undo` action.
+    pub fn undo(&mut self) {
+        self.actions.push(QueueAction::Undo);
+    }
+
+    /// Queues a `redo` action.
+    pub fn redo(&mut self) {
+        self.actions.push(QueueAction::Redo);
+    }
+
+    /// Applies the queued actions.
+    ///
+    /// # Errors
+    /// If an error occurs, it stops applying the actions and returns the error.
+    pub fn commit(self, target: &mut A::Target) -> Option<Result<A>> {
+        for action in self.actions {
+            let r = match action {
+                QueueAction::Apply(action) => Some(self.timeline.apply(target, action)),
+                QueueAction::Undo => self.timeline.undo(target),
+                QueueAction::Redo => self.timeline.redo(target),
+            };
+            match r {
+                Some(Ok(_)) => (),
+                o @ Some(Err(_)) | o @ None => return o,
+            }
+        }
+        Some(Ok(()))
+    }
+
+    /// Cancels the queued actions.
+    pub fn cancel(self) {}
+
+    /// Returns a queue.
+    pub fn queue(&mut self) -> Queue<A, F> {
+        self.timeline.queue()
+    }
+
+    /// Returns a checkpoint.
+    pub fn checkpoint(&mut self) -> Checkpoint<A, F> {
+        self.timeline.checkpoint()
+    }
+}
+
+impl<'a, A, F> From<&'a mut Timeline<A, F>> for Queue<'a, A, F> {
+    fn from(timeline: &'a mut Timeline<A, F>) -> Self {
+        Queue {
+            timeline,
+            actions: Vec::new(),
+        }
+    }
+}
+
+#[derive(Debug)]
+enum CheckpointAction<A> {
+    Apply(Option<usize>, VecDeque<Entry<A>>),
+    Undo,
+    Redo,
+}
+
+/// Wraps a timeline and gives it checkpoint functionality.
+#[derive(Debug)]
+pub struct Checkpoint<'a, A, F> {
+    timeline: &'a mut Timeline<A, F>,
+    actions: Vec<CheckpointAction<A>>,
+}
+
+impl<A: Action<Output = ()>, F: Slot> Checkpoint<'_, A, F> {
+    /// Calls the `apply` method.
+    pub fn apply(&mut self, target: &mut A::Target, action: A) -> Result<A> {
+        let saved = self.timeline.stack.saved;
+        let (_, _, tail) = self.timeline.stack.apply(target, action)?;
+        self.actions
+            .push(CheckpointAction::Apply(saved, tail.deque));
+        Ok(())
+    }
+
+    /// Calls the `undo` method.
+    pub fn undo(&mut self, target: &mut A::Target) -> Option<Result<A>> {
+        match self.timeline.undo(target) {
+            o @ Some(Ok(())) => {
+                self.actions.push(CheckpointAction::Undo);
+                o
+            }
+            o => o,
+        }
+    }
+
+    /// Calls the `redo` method.
+    pub fn redo(&mut self, target: &mut A::Target) -> Option<Result<A>> {
+        match self.timeline.redo(target) {
+            o @ Some(Ok(())) => {
+                self.actions.push(CheckpointAction::Redo);
+                o
+            }
+            o => o,
+        }
+    }
+
+    /// Commits the changes and consumes the checkpoint.
+    pub fn commit(self) {}
+
+    /// Cancels the changes and consumes the checkpoint.
+    ///
+    /// # Errors
+    /// If an error occur when canceling the changes, the error is returned
+    /// and the remaining actions are not canceled.
+    pub fn cancel(self, target: &mut A::Target) -> Option<Result<A>> {
+        for action in self.actions.into_iter().rev() {
+            match action {
+                CheckpointAction::Apply(saved, mut entries) => match self.timeline.undo(target) {
+                    Some(Ok(())) => {
+                        self.timeline.stack.entries.deque.pop_back();
+                        self.timeline.stack.entries.deque.append(&mut entries);
+                        self.timeline.stack.saved = saved;
+                    }
+                    o => return o,
+                },
+                CheckpointAction::Undo => match self.timeline.redo(target) {
+                    Some(Ok(())) => (),
+                    o => return o,
+                },
+                CheckpointAction::Redo => match self.timeline.undo(target) {
+                    Some(Ok(())) => (),
+                    o => return o,
+                },
+            };
+        }
+        Some(Ok(()))
+    }
+
+    /// Returns a queue.
+    pub fn queue(&mut self) -> Queue<A, F> {
+        self.timeline.queue()
+    }
+
+    /// Returns a checkpoint.
+    pub fn checkpoint(&mut self) -> Checkpoint<A, F> {
+        self.timeline.checkpoint()
+    }
+}
+
+impl<'a, A, F> From<&'a mut Timeline<A, F>> for Checkpoint<'a, A, F> {
+    fn from(timeline: &'a mut Timeline<A, F>) -> Self {
+        Checkpoint {
+            timeline,
+            actions: Vec::new(),
+        }
     }
 }
 
 /// Configurable display formatting for the timeline.
-#[cfg(feature = "alloc")]
-pub struct Display<'a, A, F, const LIMIT: usize> {
-    timeline: &'a Timeline<A, F, LIMIT>,
+pub struct Display<'a, A, F> {
+    timeline: &'a Timeline<A, F>,
     format: Format,
 }
 
-#[cfg(feature = "alloc")]
-impl<A, F, const LIMIT: usize> Display<'_, A, F, LIMIT> {
+impl<A, F> Display<'_, A, F> {
     /// Show colored output (on by default).
     ///
     /// Requires the `colored` feature to be enabled.
@@ -411,8 +588,7 @@ impl<A, F, const LIMIT: usize> Display<'_, A, F, LIMIT> {
     }
 }
 
-#[cfg(feature = "alloc")]
-impl<A: fmt::Display, F, const LIMIT: usize> Display<'_, A, F, LIMIT> {
+impl<A: fmt::Display, F> Display<'_, A, F> {
     fn fmt_list(&self, f: &mut fmt::Formatter, at: At, entry: Option<&Entry<A>>) -> fmt::Result {
         self.format.position(f, at, false)?;
 
@@ -427,7 +603,7 @@ impl<A: fmt::Display, F, const LIMIT: usize> Display<'_, A, F, LIMIT> {
             f,
             at,
             At::new(0, self.timeline.current()),
-            self.timeline.saved.map(|saved| At::new(0, saved)),
+            self.timeline.stack.saved.map(|saved| At::new(0, saved)),
         )?;
         if let Some(entry) = entry {
             if self.format.detailed {
@@ -443,9 +619,8 @@ impl<A: fmt::Display, F, const LIMIT: usize> Display<'_, A, F, LIMIT> {
     }
 }
 
-#[cfg(feature = "alloc")]
-impl<'a, A, F, const LIMIT: usize> From<&'a Timeline<A, F, LIMIT>> for Display<'a, A, F, LIMIT> {
-    fn from(timeline: &'a Timeline<A, F, LIMIT>) -> Self {
+impl<'a, A, F> From<&'a Timeline<A, F>> for Display<'a, A, F> {
+    fn from(timeline: &'a Timeline<A, F>) -> Self {
         Display {
             timeline,
             format: Format::default(),
@@ -453,10 +628,9 @@ impl<'a, A, F, const LIMIT: usize> From<&'a Timeline<A, F, LIMIT>> for Display<'
     }
 }
 
-#[cfg(feature = "alloc")]
-impl<A: fmt::Display, F, const LIMIT: usize> fmt::Display for Display<'_, A, F, LIMIT> {
+impl<A: fmt::Display, F> fmt::Display for Display<'_, A, F> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        for (i, entry) in self.timeline.entries.iter().enumerate().rev() {
+        for (i, entry) in self.timeline.stack.entries.deque.iter().enumerate().rev() {
             let at = At::new(0, i + 1);
             self.fmt_list(f, at, Some(entry))?;
         }
@@ -464,37 +638,286 @@ impl<A: fmt::Display, F, const LIMIT: usize> fmt::Display for Display<'_, A, F, 
     }
 }
 
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug)]
+pub(crate) struct LimitDeque<T> {
+    pub deque: VecDeque<Entry<T>>,
+    pub limit: NonZeroUsize,
+}
+
+impl<T> Entries for LimitDeque<T> {
+    type Item = T;
+
+    fn limit(&self) -> usize {
+        self.limit.get()
+    }
+
+    fn len(&self) -> usize {
+        self.deque.len()
+    }
+
+    fn back_mut(&mut self) -> Option<&mut Entry<T>> {
+        self.deque.back_mut()
+    }
+
+    fn push_back(&mut self, t: Entry<T>) {
+        self.deque.push_back(t)
+    }
+
+    fn pop_front(&mut self) -> Option<Entry<T>> {
+        self.deque.pop_front()
+    }
+
+    fn pop_back(&mut self) -> Option<Entry<T>> {
+        self.deque.pop_back()
+    }
+
+    fn split_off(&mut self, at: usize) -> Self {
+        LimitDeque {
+            deque: self.deque.split_off(at),
+            limit: self.limit,
+        }
+    }
+
+    fn clear(&mut self) {
+        self.deque.clear();
+    }
+}
+
+impl<T> Index<usize> for LimitDeque<T> {
+    type Output = Entry<T>;
+
+    fn index(&self, index: usize) -> &Self::Output {
+        self.deque.index(index)
+    }
+}
+
+impl<T> IndexMut<usize> for LimitDeque<T> {
+    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
+        self.deque.index_mut(index)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use crate::*;
-    use arrayvec::ArrayString;
+    use alloc::string::String;
+
+    enum Edit {
+        Add(Add),
+        Del(Del),
+    }
+
+    impl Action for Edit {
+        type Target = String;
+        type Output = ();
+        type Error = &'static str;
+
+        fn apply(&mut self, s: &mut String) -> Result<Add> {
+            match self {
+                Edit::Add(add) => add.apply(s),
+                Edit::Del(del) => del.apply(s),
+            }
+        }
+
+        fn undo(&mut self, s: &mut String) -> Result<Add> {
+            match self {
+                Edit::Add(add) => add.undo(s),
+                Edit::Del(del) => del.undo(s),
+            }
+        }
+
+        fn merge(&mut self, edit: Self) -> Merged<Self>
+        where
+            Self: Sized,
+        {
+            match (self, edit) {
+                (Edit::Add(_), Edit::Del(_)) => Merged::Annul,
+                (Edit::Del(Del(Some(a))), Edit::Add(Add(b))) if a == &b => Merged::Annul,
+                (_, edit) => Merged::No(edit),
+            }
+        }
+    }
 
     struct Add(char);
 
     impl Action for Add {
-        type Target = ArrayString<64>;
+        type Target = String;
         type Output = ();
         type Error = &'static str;
 
-        fn apply(&mut self, s: &mut ArrayString<64>) -> Result<Add> {
+        fn apply(&mut self, s: &mut String) -> Result<Add> {
             s.push(self.0);
             Ok(())
         }
 
-        fn undo(&mut self, s: &mut ArrayString<64>) -> Result<Add> {
+        fn undo(&mut self, s: &mut String) -> Result<Add> {
             self.0 = s.pop().ok_or("s is empty")?;
             Ok(())
         }
     }
 
-    #[test]
-    fn limit() {
-        let mut target = ArrayString::new();
-        let mut timeline = Timeline::<_, _, 32>::new();
-        for i in 64..128 {
-            timeline.apply(&mut target, Add(char::from(i))).unwrap();
+    #[derive(Default)]
+    struct Del(Option<char>);
+
+    impl Action for Del {
+        type Target = String;
+        type Output = ();
+        type Error = &'static str;
+
+        fn apply(&mut self, s: &mut String) -> Result<Add> {
+            self.0 = s.pop();
+            Ok(())
         }
-        assert_eq!(target.len(), 64);
-        assert_eq!(timeline.len(), 32);
+
+        fn undo(&mut self, s: &mut String) -> Result<Add> {
+            let ch = self.0.ok_or("s is empty")?;
+            s.push(ch);
+            Ok(())
+        }
+    }
+
+    #[test]
+    fn go_to() {
+        let mut target = String::new();
+        let mut timeline = Timeline::new();
+        timeline.apply(&mut target, Add('a')).unwrap();
+        timeline.apply(&mut target, Add('b')).unwrap();
+        timeline.apply(&mut target, Add('c')).unwrap();
+        timeline.apply(&mut target, Add('d')).unwrap();
+        timeline.apply(&mut target, Add('e')).unwrap();
+
+        timeline.go_to(&mut target, 0).unwrap().unwrap();
+        assert_eq!(timeline.current(), 0);
+        assert_eq!(target, "");
+        timeline.go_to(&mut target, 5).unwrap().unwrap();
+        assert_eq!(timeline.current(), 5);
+        assert_eq!(target, "abcde");
+        timeline.go_to(&mut target, 1).unwrap().unwrap();
+        assert_eq!(timeline.current(), 1);
+        assert_eq!(target, "a");
+        timeline.go_to(&mut target, 4).unwrap().unwrap();
+        assert_eq!(timeline.current(), 4);
+        assert_eq!(target, "abcd");
+        timeline.go_to(&mut target, 2).unwrap().unwrap();
+        assert_eq!(timeline.current(), 2);
+        assert_eq!(target, "ab");
+        timeline.go_to(&mut target, 3).unwrap().unwrap();
+        assert_eq!(timeline.current(), 3);
+        assert_eq!(target, "abc");
+        assert!(timeline.go_to(&mut target, 6).is_none());
+        assert_eq!(timeline.current(), 3);
+    }
+
+    #[test]
+    fn queue_commit() {
+        let mut target = String::new();
+        let mut timeline = Timeline::new();
+        let mut q1 = timeline.queue();
+        q1.redo();
+        q1.redo();
+        q1.redo();
+        let mut q2 = q1.queue();
+        q2.undo();
+        q2.undo();
+        q2.undo();
+        let mut q3 = q2.queue();
+        q3.apply(Add('a'));
+        q3.apply(Add('b'));
+        q3.apply(Add('c'));
+        assert_eq!(target, "");
+        q3.commit(&mut target).unwrap().unwrap();
+        assert_eq!(target, "abc");
+        q2.commit(&mut target).unwrap().unwrap();
+        assert_eq!(target, "");
+        q1.commit(&mut target).unwrap().unwrap();
+        assert_eq!(target, "abc");
+    }
+
+    #[test]
+    fn checkpoint_commit() {
+        let mut target = String::new();
+        let mut timeline = Timeline::new();
+        let mut cp1 = timeline.checkpoint();
+        cp1.apply(&mut target, Add('a')).unwrap();
+        cp1.apply(&mut target, Add('b')).unwrap();
+        cp1.apply(&mut target, Add('c')).unwrap();
+        assert_eq!(target, "abc");
+        let mut cp2 = cp1.checkpoint();
+        cp2.apply(&mut target, Add('d')).unwrap();
+        cp2.apply(&mut target, Add('e')).unwrap();
+        cp2.apply(&mut target, Add('f')).unwrap();
+        assert_eq!(target, "abcdef");
+        let mut cp3 = cp2.checkpoint();
+        cp3.apply(&mut target, Add('g')).unwrap();
+        cp3.apply(&mut target, Add('h')).unwrap();
+        cp3.apply(&mut target, Add('i')).unwrap();
+        assert_eq!(target, "abcdefghi");
+        cp3.commit();
+        cp2.commit();
+        cp1.commit();
+        assert_eq!(target, "abcdefghi");
+    }
+
+    #[test]
+    fn checkpoint_cancel() {
+        let mut target = String::new();
+        let mut timeline = Timeline::new();
+        let mut cp1 = timeline.checkpoint();
+        cp1.apply(&mut target, Add('a')).unwrap();
+        cp1.apply(&mut target, Add('b')).unwrap();
+        cp1.apply(&mut target, Add('c')).unwrap();
+        let mut cp2 = cp1.checkpoint();
+        cp2.apply(&mut target, Add('d')).unwrap();
+        cp2.apply(&mut target, Add('e')).unwrap();
+        cp2.apply(&mut target, Add('f')).unwrap();
+        let mut cp3 = cp2.checkpoint();
+        cp3.apply(&mut target, Add('g')).unwrap();
+        cp3.apply(&mut target, Add('h')).unwrap();
+        cp3.apply(&mut target, Add('i')).unwrap();
+        assert_eq!(target, "abcdefghi");
+        cp3.cancel(&mut target).unwrap().unwrap();
+        assert_eq!(target, "abcdef");
+        cp2.cancel(&mut target).unwrap().unwrap();
+        assert_eq!(target, "abc");
+        cp1.cancel(&mut target).unwrap().unwrap();
+        assert_eq!(target, "");
+    }
+
+    #[test]
+    fn checkpoint_saved() {
+        let mut target = String::new();
+        let mut timeline = Timeline::new();
+        timeline.apply(&mut target, Add('a')).unwrap();
+        timeline.apply(&mut target, Add('b')).unwrap();
+        timeline.apply(&mut target, Add('c')).unwrap();
+        timeline.set_saved(true);
+        timeline.undo(&mut target).unwrap().unwrap();
+        timeline.undo(&mut target).unwrap().unwrap();
+        timeline.undo(&mut target).unwrap().unwrap();
+        let mut cp = timeline.checkpoint();
+        cp.apply(&mut target, Add('d')).unwrap();
+        cp.apply(&mut target, Add('e')).unwrap();
+        cp.apply(&mut target, Add('f')).unwrap();
+        assert_eq!(target, "def");
+        cp.cancel(&mut target).unwrap().unwrap();
+        assert_eq!(target, "");
+        timeline.redo(&mut target).unwrap().unwrap();
+        timeline.redo(&mut target).unwrap().unwrap();
+        timeline.redo(&mut target).unwrap().unwrap();
+        assert!(timeline.is_saved());
+        assert_eq!(target, "abc");
+    }
+
+    #[test]
+    fn annul() {
+        let mut target = String::new();
+        let mut timeline = Timeline::new();
+        timeline.apply(&mut target, Edit::Add(Add('a'))).unwrap();
+        timeline
+            .apply(&mut target, Edit::Del(Del::default()))
+            .unwrap();
+        timeline.apply(&mut target, Edit::Add(Add('b'))).unwrap();
+        assert_eq!(timeline.len(), 1);
     }
 }
